test(stella-cli): split plugin_cmd/tests.rs at the package seam — it crossed the 1500-line ceiling
The file reached 1511 lines on main through two PRs that each grew it
correctly (#4404, #4416); the ratchet judges a change, not the tree, so the
crossing was caught only by the post-merge canary (#4433). The contributed
tools/skills/records tests — the block under the 'A plugin is a package'
divider — move to plugin_cmd/tests/contributions.rs as a child module; the
shared fixtures stay in the parent and reach it through super::*. Pure move:
27 tests before and after, names unchanged. No baseline entry.

Closes #4440
Closes #4433
